- Probably fixed the Segmentation Faults we've been having.   Description: A player quits and it's session is freed and set to NULL,        but the char server already sent a packet with player information        (registers, storage, ...). If a message is sent in consequence of        updating that info, a segmentation fault happens because        session[sd->fd] is already NULL.   Fix: make shure the session of the target player is valid before        processing the rest of the char server's message. - Some minor cleanups.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@9366 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/socket.h
@@ -103,9 +103,9 @@ enum SessionType {
 struct socket_data{
 	unsigned char eof;
 	unsigned char *rdata, *wdata;
-	unsigned int max_rdata, max_wdata;
-	unsigned int rdata_size, wdata_size;
-	int rdata_pos;
+	size_t max_rdata, max_wdata;
+	size_t rdata_size, wdata_size;
+	size_t rdata_pos;
 	time_t rdata_tick;
 	struct sockaddr_in client_addr;
 	int (*func_recv)(int);
